Rely on default `files` list for hex package (#98)
The initial goal for this change was to incorporate the CHANGELOG into
the hex package so that it would be available in the diff tool on hex.pm
or the command line (i.e. `mix hex.package diff thousand_island 1.1.0..1.2.0`).

Per discussion in mtrudel/bandit#265 we'll accomplish the above by
removing the `files` configuration and relying on the default list of
files provided by hex (see the linked issue for details).

This change will also have the effect of removing test files from future
releases.